# Conflicts:
#	sormas-api/src/main/java/de/symeda/sormas/api/i18n/Captions.java
#	sormas-api/src/main/java/de/symeda/sormas/api/i18n/Strings.java
#	sormas-api/src/main/java/de/symeda/sormas/api/sample/SampleCriteria.java
#	sormas-api/src/main/java/de/symeda/sormas/api/sample/SampleDto.java
#	sormas-api/src/main/java/de/symeda/sormas/api/sample/SampleIndexDto.java
#	sormas-api/src/main/resources/enum.properties
#	sormas-api/src/main/resources/strings.properties
#	sormas-app/app/src/main/java/de/symeda/sormas/app/backend/common/DatabaseHelper.java
#	sormas-app/app/src/main/java/de/symeda/sormas/app/epidata/EpidemiologicalDataEditFragment.java
#	sormas-app/app/src/main/res/layout/fragment_sample_read_layout.xml
#	sormas-backend/src/main/java/de/symeda/sormas/backend/caze/CaseService.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/sample/SampleFacadeEjb.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/sample/SampleJoins.java
#	sormas-backend/src/main/java/de/symeda/sormas/backend/sample/SampleService.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/caze/CaseController.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/caze/importer/CaseImporter.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/PathogenTestController.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/PathogenTestForm.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/SampleController.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/SampleGrid.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/SampleGridComponent.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/SampleGridFilterForm.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/samples/SamplesView.java

+++ sormas-ui/src/main/java/de/symeda/sormas/ui/campaign/CampaignController.java
@@ -15,6 +15,9 @@
 
 package de.symeda.sormas.ui.campaign;
 
+import static com.vaadin.v7.data.Validator.InvalidValueException;
+import static de.symeda.sormas.api.campaign.data.CampaignFormDataCriteria.CAMPAIGN;
+
 import com.vaadin.ui.Alignment;
 import com.vaadin.ui.Button;
 import com.vaadin.ui.Label;
@@ -23,14 +26,17 @@ import com.vaadin.ui.Notification.Type;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.Window;
 import com.vaadin.ui.themes.ValoTheme;
+
 import de.symeda.sormas.api.FacadeProvider;
 import de.symeda.sormas.api.campaign.CampaignDto;
+import de.symeda.sormas.api.campaign.CampaignReferenceDto;
 import de.symeda.sormas.api.campaign.data.CampaignFormDataDto;
-import de.symeda.sormas.api.campaign.form.CampaignFormReferenceDto;
+import de.symeda.sormas.api.campaign.form.CampaignFormMetaReferenceDto;
 import de.symeda.sormas.api.i18n.Captions;
 import de.symeda.sormas.api.i18n.I18nProperties;
 import de.symeda.sormas.api.i18n.Strings;
 import de.symeda.sormas.api.i18n.Validations;
+import de.symeda.sormas.api.user.UserDto;
 import de.symeda.sormas.api.user.UserRight;
 import de.symeda.sormas.ui.SormasUI;
 import de.symeda.sormas.ui.UserProvider;
@@ -38,11 +44,11 @@ import de.symeda.sormas.ui.campaign.campaigndata.CampaignDataView;
 import de.symeda.sormas.ui.campaign.campaigndata.CampaignFormDataEditForm;
 import de.symeda.sormas.ui.campaign.campaigndata.CampaignFormDataView;
 import de.symeda.sormas.ui.campaign.campaigns.CampaignEditForm;
+import de.symeda.sormas.ui.campaign.campaigns.CampaignView;
+import de.symeda.sormas.ui.campaign.campaigns.CampaignsView;
 import de.symeda.sormas.ui.utils.ButtonHelper;
 import de.symeda.sormas.ui.utils.CommitDiscardWrapperComponent;
 import de.symeda.sormas.ui.utils.VaadinUiUtil;
-
-import static com.vaadin.v7.data.Validator.InvalidValueException;
 
 public class CampaignController {
 
@@ -87,15 +93,16 @@ public class CampaignController {
 		VaadinUiUtil.showModalPopupWindow(campaignComponent, heading);
 	}
 
-	public void createCampaignDataForm(CampaignFormReferenceDto campaignForm) {
+	public void createCampaignDataForm(CampaignReferenceDto campaign, CampaignFormMetaReferenceDto campaignForm) {
 		Window window = VaadinUiUtil.createPopupWindow();
 
-		CommitDiscardWrapperComponent<CampaignFormDataEditForm> component = getCampaignFormDataComponent(null, campaignForm, false, false, () -> {
-			window.close();
-			SormasUI.refreshView();
-			Notification
-				.show(String.format(I18nProperties.getString(Strings.messageCampaignFormSaved), campaignForm.toString()), Type.TRAY_NOTIFICATION);
-		}, window::close);
+		CommitDiscardWrapperComponent<CampaignFormDataEditForm> component =
+			getCampaignFormDataComponent(null, campaign, campaignForm, false, false, () -> {
+				window.close();
+				SormasUI.refreshView();
+				Notification
+					.show(String.format(I18nProperties.getString(Strings.messageCampaignFormSaved), campaignForm.toString()), Type.TRAY_NOTIFICATION);
+			}, window::close);
 
 		window.setCaption(String.format(I18nProperties.getString(Strings.headingCreateCampaignDataForm), campaignForm.toString()));
 		window.setContent(component);
@@ -145,44 +152,76 @@ public class CampaignController {
 
 	public CommitDiscardWrapperComponent<CampaignEditForm> getCampaignComponent(CampaignDto campaignDto, Runnable callback) {
 
-		CampaignEditForm campaignEditForm = new CampaignEditForm(campaignDto == null);
-
-		final CommitDiscardWrapperComponent<CampaignEditForm> view =
-			new CommitDiscardWrapperComponent<CampaignEditForm>(campaignEditForm, campaignEditForm.getFieldGroup());
-
+		CampaignEditForm campaignEditForm = new CampaignEditForm(campaignDto);
 		if (campaignDto == null) {
 			campaignDto = CampaignDto.build();
 			campaignDto.setCreatingUser(UserProvider.getCurrent().getUserReference());
 		}
 		campaignEditForm.setValue(campaignDto);
 
-		view.addCommitListener(() -> {
+		final CommitDiscardWrapperComponent<CampaignEditForm> campaignComponent =
+			new CommitDiscardWrapperComponent<CampaignEditForm>(campaignEditForm, UserProvider.getCurrent().hasUserRight(UserRight.CAMPAIGN_EDIT), campaignEditForm.getFieldGroup()) {
+
+				@Override
+				public void discard() {
+					super.discard();
+					campaignEditForm.discard();
+				}
+			};
+
+		if (UserProvider.getCurrent().hasUserRight(UserRight.CAMPAIGN_DELETE)) {
+			CampaignDto finalCampaignDto = campaignDto;
+			campaignComponent.addDeleteListener(() -> {
+				FacadeProvider.getCampaignFacade().deleteCampaign(finalCampaignDto.getUuid());
+				UI.getCurrent().getNavigator().navigateTo(CampaignsView.VIEW_NAME);
+			}, I18nProperties.getString(Strings.entityCampaign));
+		}
+
+		// Initialize 'Archive' button
+		if (UserProvider.getCurrent().hasUserRight(UserRight.CAMPAIGN_ARCHIVE)) {
+			final String campaignUuid = campaignDto.getUuid();
+			boolean archived = FacadeProvider.getCampaignFacade().isArchived(campaignUuid);
+			Button archiveCampaignButton = ButtonHelper.createButton(archived ? Captions.actionDearchive : Captions.actionArchive, e -> {
+				campaignComponent.commit();
+				archiveOrDearchiveCampaign(campaignUuid, !archived);
+			}, ValoTheme.BUTTON_LINK);
+
+			campaignComponent.getButtonsPanel().addComponentAsFirst(archiveCampaignButton);
+			campaignComponent.getButtonsPanel().setComponentAlignment(archiveCampaignButton, Alignment.BOTTOM_LEFT);
+		}
+
+		campaignComponent.addCommitListener(() -> {
 			if (!campaignEditForm.getFieldGroup().isModified()) {
 				CampaignDto dto = campaignEditForm.getValue();
 				FacadeProvider.getCampaignFacade().saveCampaign(dto);
+				SormasUI.refreshView();
 				callback.run();
 			}
 		});
 
-		return view;
+		return campaignComponent;
 	}
 
 	public CommitDiscardWrapperComponent<CampaignFormDataEditForm> getCampaignFormDataComponent(
 		CampaignFormDataDto campaignFormData,
-		CampaignFormReferenceDto campaignForm,
+		CampaignReferenceDto campaign,
+		CampaignFormMetaReferenceDto campaignForm,
 		boolean revertFormOnDiscard,
 		boolean showDeleteButton,
 		Runnable commitCallback,
 		Runnable discardCallback) {
+
 		CampaignFormDataEditForm form = new CampaignFormDataEditForm(campaignFormData == null);
-
-		final CommitDiscardWrapperComponent<CampaignFormDataEditForm> component = new CommitDiscardWrapperComponent<>(form, form.getFieldGroup());
-
 		if (campaignFormData == null) {
-			campaignFormData = CampaignFormDataDto.build(null, campaignForm, null, null, null);
+
+			final UserDto currentUser = UserProvider.getCurrent().getUser();
+			campaignFormData =
+				CampaignFormDataDto.build(campaign, campaignForm, currentUser.getRegion(), currentUser.getDistrict(), currentUser.getCommunity());
+			campaignFormData.setCreatingUser(UserProvider.getCurrent().getUserReference());
 		}
 		form.setValue(campaignFormData);
-		final String campaignFormDataUuid = campaignFormData.getUuid();
+
+		final CommitDiscardWrapperComponent<CampaignFormDataEditForm> component = new CommitDiscardWrapperComponent<>(form, form.getFieldGroup());
 
 		component.addCommitListener(() -> {
 			if (!form.getFieldGroup().isModified()) {
@@ -197,9 +236,13 @@ public class CampaignController {
 				FacadeProvider.getCampaignFormDataFacade().saveCampaignFormData(formData);
 				if (commitCallback != null) {
 					commitCallback.run();
+					UI.getCurrent().getNavigator().navigateTo(CampaignDataView.VIEW_NAME + "/?" + CAMPAIGN + "=" + campaign.getUuid());
 				}
 			}
 		});
+
+		component.addDiscardListener(
+			() -> UI.getCurrent().getNavigator().navigateTo(CampaignDataView.VIEW_NAME + "/?" + CAMPAIGN + "=" + campaign.getUuid()));
 
 		if (revertFormOnDiscard) {
 			component.addDiscardListener(form::resetFormValues);
@@ -209,7 +252,9 @@ public class CampaignController {
 			component.addDiscardListener(discardCallback::run);
 		}
 
-		if (showDeleteButton) {
+		if (showDeleteButton && UserProvider.getCurrent().hasUserRight(UserRight.CAMPAIGN_DELETE)) {
+			String campaignFormDataUuid = campaignFormData.getUuid();
+
 			component.addDeleteListener(() -> {
 				FacadeProvider.getCampaignFormDataFacade().deleteCampaignFormData(campaignFormDataUuid);
 				UI.getCurrent().getNavigator().navigateTo(CampaignFormDataView.VIEW_NAME);
@@ -221,6 +266,11 @@ public class CampaignController {
 
 	private CampaignDto getCampaign(String uuid) {
 		return FacadeProvider.getCampaignFacade().getByUuid(uuid);
+	}
+
+	public void navigateToCampaign(String uuid) {
+		String navigationState = CampaignView.VIEW_NAME + "/" + uuid;
+		SormasUI.get().getNavigator().navigateTo(navigationState);
 	}
 
 	public void navigateToFormDataView(String uuid) {
